Finished writing the short scripts to transmit the potentiometer value from one feather to another.
The transmitter will read and transmit the pot value (between 0 and 255).
The reciever will then read that value in, print it to serial along with diagnostic values, and then set the PWM pin 5 to the pot value.

This way if you set up an LED for both feathers, they should brighten and dim togeather.