serial_tx/serial_tx.ino: Transmits whatever is being sent to the serial port that is connected to it via radio.
serial_rx/serial_rx.ino: Listens to the radio and sends whatever it hears via serial port.

send_data.py: This script (with command line arguments) sends a test data set to a serial port specified by the user.

listen_port.py: Listens to a port supplied with a command line argument. When data is recieved, it is decoded and saved to a test_file.txt.